Refactor : HttpRequestParser와 HttpRequestFactory 통합
HttpRequestParser의 기능이 HttpRequestFactory 안에 있는 것이 HttpRequestFactory객체의 본질을 다한다는 생각이 들어 내부로 편입
